llm-rag-vector-db/requirements.txt
numpy
scikit-learn
faiss-cpu
pandas
jsonschema
torch
transformers